removed unused/redundant code
