+++ package.json
@@ -52,6 +52,7 @@
     "@lumino/datastore": "^0.10.3",
     "@lumino/dragdrop": "^1.6.2",
     "@lumino/widgets": "^1.13.2",
+    "eslint-plugin-react": "^7.21.1",
     "husky": "^4.2.5",
     "lint-staged": "^10.2.11"
   },
