--- src/dialog.ts
@@ -1,25 +1,0 @@
-import { Dialog } from '@jupyterlab/apputils';
-import { Dashboard } from './dashboard';
-
-export function unsaveDialog(dashboard: Dashboard): Dialog<unknown> {
-  const dialog = new Dialog({
-    title: 'Close without saving?',
-    body: `"${dashboard.model.name}.dashboard" has unsaved changes, close without saving?`,
-    buttons: [
-      Dialog.cancelButton(),
-      Dialog.okButton({
-        label: 'OK',
-      }),
-    ],
-  });
-  return dialog;
-}
-
-export function saveDialog(filename: string): Dialog<unknown> {
-  const dialog = new Dialog({
-    title: 'Dashboard saved',
-    body: 'All changes to "' + filename + ' have been saved.',
-    buttons: [Dialog.okButton()],
-  });
-  return dialog;
-}
