--- src/fullscreen.ts
@@ -1,21 +1,0 @@
-export function openfullscreen(node: HTMLElement): void {
-  // Trigger fullscreen
-  const docElmWithBrowsersFullScreenFunctions = node as HTMLElement & {
-    mozRequestFullScreen(): Promise<void>;
-    webkitRequestFullscreen(): Promise<void>;
-    msRequestFullscreen(): Promise<void>;
-  };
-
-  if (docElmWithBrowsersFullScreenFunctions.requestFullscreen) {
-    docElmWithBrowsersFullScreenFunctions.requestFullscreen();
-  } else if (docElmWithBrowsersFullScreenFunctions.mozRequestFullScreen) {
-    /* Firefox */
-    docElmWithBrowsersFullScreenFunctions.mozRequestFullScreen();
-  } else if (docElmWithBrowsersFullScreenFunctions.webkitRequestFullscreen) {
-    /* Chrome, Safari and Opera */
-    docElmWithBrowsersFullScreenFunctions.webkitRequestFullscreen();
-  } else if (docElmWithBrowsersFullScreenFunctions.msRequestFullscreen) {
-    /* IE/Edge */
-    docElmWithBrowsersFullScreenFunctions.msRequestFullscreen();
-  }
-}
