+++ src/widget.ts
@@ -12,8 +12,6 @@ import { Message } from '@lumino/messaging';
 
 import { Drag } from './drag';
 
-import { shouldStartDrag } from './widgetdragutils';
-
 import { Dashboard } from './dashboard';
 
 import {
@@ -322,7 +320,10 @@ export class DashboardWidget extends Widget {
     const data = this._clickData;
     const { clientX, clientY } = event;
 
-    if (data && shouldStartDrag(data.pressX, data.pressY, clientX, clientY)) {
+    if (
+      data &&
+      Private.shouldStartDrag(data.pressX, data.pressY, clientX, clientY)
+    ) {
       void this._startDrag(data.target, clientX, clientY);
     }
   }
@@ -731,6 +732,33 @@ export namespace DashboardWidget {
    * How many pixels to adjust a widget by using the arrow keys.
    */
   export const BUMP_DISTANCE = 10;
+}
+
+/**
+ * A namespace for private functionality.
+ */
+namespace Private {
+  const DRAG_THRESHOLD = 5;
+
+  /**
+   * Detect if a drag event should be started. This is down if the
+   * mouse is moved beyond a certain distance (DRAG_THRESHOLD).
+   *
+   * @param prevX - X Coordinate of the mouse pointer during the mousedown event
+   * @param prevY - Y Coordinate of the mouse pointer during the mousedown event
+   * @param nextX - Current X Coordinate of the mouse pointer
+   * @param nextY - Current Y Coordinate of the mouse pointer
+   */
+  export function shouldStartDrag(
+    prevX: number,
+    prevY: number,
+    nextX: number,
+    nextY: number
+  ): boolean {
+    const dx = Math.abs(nextX - prevX);
+    const dy = Math.abs(nextY - prevY);
+    return dx >= DRAG_THRESHOLD || dy >= DRAG_THRESHOLD;
+  }
 }
 
 export default DashboardWidget;
